vm: refactor runCallOpts and runCodeOpts to types and re-add their docs. Order imports in evm.ts
